@@ -21,10 +21,9 @@ namespace PlaylistManager.UI
     {
         private readonly StandardLevelDetailViewController standardLevelDetailViewController;
         private readonly PopupModalsController popupModalsController;
-        private readonly IVRPlatformHelper platformHelper;
 
         private BeatSaberPlaylistsLib.PlaylistManager parentManager;
-        private BeatSaberPlaylistsLib.PlaylistManager[] childManagers;
+        private List<BeatSaberPlaylistsLib.PlaylistManager> childManagers;
         private List<BeatSaberPlaylistsLib.Types.IPlaylist> childPlaylists;
 
         private readonly Sprite folderIcon;
@@ -32,10 +31,10 @@ namespace PlaylistManager.UI
         public event PropertyChangedEventHandler PropertyChanged;
 
         [UIComponent("list")]
-        public CustomListTableData customListTableData;
+        public CustomListTableData playlistTableData;
 
-        [UIComponent("scroll-view")]
-        private ScrollView bsmlScrollView;
+        [UIComponent("dropdown-options")]
+        public CustomListTableData dropdownTableData;
 
         [UIComponent("highlight-checkbox")]
         private readonly RectTransform highlightCheckboxTransform;
@@ -45,14 +44,21 @@ namespace PlaylistManager.UI
 
         private Vector3 modalPosition;
 
+        [UIComponent("create-dropdown")]
+        private ModalView createModal;
+
+        [UIComponent("create-dropdown")]
+        private readonly RectTransform createModalTransform;
+
+        private Vector3 createModalPosition;
+
         [UIParams]
         private readonly BSMLParserParams parserParams;
 
-        public AddPlaylistModalController(StandardLevelDetailViewController standardLevelDetailViewController, PopupModalsController popupModalsController, IVRPlatformHelper platformHelper)
+        public AddPlaylistModalController(StandardLevelDetailViewController standardLevelDetailViewController, PopupModalsController popupModalsController)
         {
             this.standardLevelDetailViewController = standardLevelDetailViewController;
             this.popupModalsController = popupModalsController;
-            this.platformHelper = platformHelper;
             folderIcon = BeatSaberMarkupLanguage.Utilities.FindSpriteInAssembly("PlaylistManager.Icons.FolderIcon.png");
             parsed = false;
         }
@@ -62,9 +68,11 @@ namespace PlaylistManager.UI
             if (!parsed)
             {
                 BSMLParser.instance.Parse(BeatSaberMarkupLanguage.Utilities.GetResourceContent(Assembly.GetExecutingAssembly(), "PlaylistManager.UI.Views.AddPlaylistModal.bsml"), standardLevelDetailViewController.transform.Find("LevelDetail").gameObject, this);
-                modalPosition = modalTransform.position; // Position can change if SongBrowser is clicked while modal is opened so storing here
+                modalPosition = modalTransform.localPosition;
+                createModalPosition = createModalTransform.localPosition;
             }
-            modalTransform.position = modalPosition; // Reset position
+            modalTransform.localPosition = modalPosition; // Reset position
+            createModalTransform.localPosition = createModalPosition;
         }
 
         [UIAction("#post-parse")]
@@ -73,9 +81,10 @@ namespace PlaylistManager.UI
             parsed = true;
             highlightCheckboxTransform.transform.localScale *= 0.5f;
 
-            ScrollView scrollView = customListTableData.tableView.GetComponent<ScrollView>();
-            Accessors.PlatformHelperAccessor(ref scrollView) = platformHelper;
-            Utils.TransferScrollBar(bsmlScrollView, scrollView);
+            Accessors.AnimateCanvasAccessor(ref createModal) = false;
+            dropdownTableData.data.Add(new CustomCellInfo("Playlist"));
+            dropdownTableData.data.Add(new CustomCellInfo("Folder"));
+            dropdownTableData.tableView.ReloadData();
         }
 
         #region Show Playlists
@@ -90,16 +99,16 @@ namespace PlaylistManager.UI
 
         internal void ShowPlaylistsForManager(BeatSaberPlaylistsLib.PlaylistManager parentManager)
         {
-            customListTableData.data.Clear();
+            playlistTableData.data.Clear();
 
             this.parentManager = parentManager;
-            childManagers = parentManager.GetChildManagers().ToArray();
+            childManagers = parentManager.GetChildManagers().ToList();
             var childPlaylists = parentManager.GetAllPlaylists(false).Where(playlist => !playlist.ReadOnly);
             this.childPlaylists = childPlaylists.ToList();
 
             foreach (BeatSaberPlaylistsLib.PlaylistManager playlistManager in childManagers)
             {
-                customListTableData.data.Add(new CustomCellInfo(Path.GetFileName(playlistManager.PlaylistPath), "Folder", folderIcon));
+                playlistTableData.data.Add(new CustomCellInfo(Path.GetFileName(playlistManager.PlaylistPath), "Folder", folderIcon));
             }
             foreach (BeatSaberPlaylistsLib.Types.IPlaylist playlist in childPlaylists)
             {
@@ -114,8 +123,7 @@ namespace PlaylistManager.UI
                     ShowPlaylist(playlist);
                 }
             }
-            customListTableData.tableView.ReloadData();
-            customListTableData.tableView.ScrollToCellWithIdx(0, TableView.ScrollPositionType.Beginning, false);
+            playlistTableData.tableView.ReloadData();
 
             PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(nameof(BackActive)));
             PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(nameof(FolderText)));
@@ -129,7 +137,7 @@ namespace PlaylistManager.UI
                 {
                     ShowPlaylist((BeatSaberPlaylistsLib.Types.IPlaylist)deferredSpriteLoadPlaylist);
                 }
-                customListTableData.tableView.ReloadDataKeepingPosition();
+                playlistTableData.tableView.ReloadDataKeepingPosition();
                 (deferredSpriteLoadPlaylist).SpriteLoaded -= DeferredSpriteLoadPlaylist_SpriteLoaded;
             }
         }
@@ -146,34 +154,36 @@ namespace PlaylistManager.UI
                 }
                 subName += " (contains song)";
             }
-            customListTableData.data.Add(new CustomCellInfo(playlist.collectionName, subName, playlist.coverImage));
+            playlistTableData.data.Add(new CustomCellInfo(playlist.collectionName, subName, playlist.coverImage));
         }
 
         [UIAction("select-cell")]
         private void OnCellSelect(TableView tableView, int index)
         {
-            customListTableData.tableView.ClearSelection();
+            playlistTableData.tableView.ClearSelection();
             // Folder Selected
-            if (index < childManagers.Length)
+            if (index < childManagers.Count)
             {
                 ShowPlaylistsForManager(childManagers[index]);
             }
             else
             {
-                index -= childManagers.Length;
+                index -= childManagers.Count;
                 var selectedPlaylist = childPlaylists[index];
+                IPlaylistSong playlistSong;
                 if (HighlightDifficulty)
                 {
-                    selectedPlaylist.Add(standardLevelDetailViewController.selectedDifficultyBeatmap);
+                    playlistSong = selectedPlaylist.Add(standardLevelDetailViewController.selectedDifficultyBeatmap);
                 }
                 else
                 {
-                    selectedPlaylist.Add(standardLevelDetailViewController.selectedDifficultyBeatmap.level);
+                    playlistSong = selectedPlaylist.Add(standardLevelDetailViewController.selectedDifficultyBeatmap.level);
                 }
                 try
                 {
                     parentManager.StorePlaylist(selectedPlaylist);
                     popupModalsController.ShowOkModal(modalTransform, string.Format("Song successfully added to {0}", selectedPlaylist.collectionName), null, animateParentCanvas: false);
+                    Events.RaisePlaylistSongAdded(playlistSong, selectedPlaylist);
                 }
                 catch (Exception e)
                 {
@@ -220,39 +230,55 @@ namespace PlaylistManager.UI
 
         #region Create Playlist
 
-        [UIAction("open-keyboard")]
-        private void OpenKeyboard()
+        [UIAction("select-option")]
+        private void OnOptionSelect(TableView tableView, int index)
         {
-            popupModalsController.ShowKeyboard(modalTransform, CreatePlaylist, animateParentCanvas: false);
+            popupModalsController.ShowKeyboard(modalTransform, index == 0 ? CreatePlaylist : CreateFolder, animateParentCanvas: false);
+            tableView.ClearSelection();
+            parserParams.EmitEvent("close-dropdown");
         }
 
         private void CreatePlaylist(string playlistName)
         {
-            BeatSaberPlaylistsLib.Types.IPlaylist playlist;
             if (string.IsNullOrWhiteSpace(playlistName))
             {
                 return;
             }
-            if (!PluginConfig.Instance.DefaultImageDisabled)
-            {
-                playlist = PlaylistLibUtils.CreatePlaylist(playlistName, PluginConfig.Instance.AuthorName, parentManager);
-            }
-            else
-            {
-                playlist = PlaylistLibUtils.CreatePlaylist(playlistName, PluginConfig.Instance.AuthorName, parentManager, false);
-            }
+
+            BeatSaberPlaylistsLib.Types.IPlaylist playlist = PlaylistLibUtils.CreatePlaylistWithConfig(playlistName, parentManager);
+
             if (playlist is IDeferredSpriteLoad deferredSpriteLoadPlaylist && !deferredSpriteLoadPlaylist.SpriteWasLoaded)
             {
-                _ = playlist.coverImage;
                 deferredSpriteLoadPlaylist.SpriteLoaded -= DeferredSpriteLoadPlaylist_SpriteLoaded;
                 deferredSpriteLoadPlaylist.SpriteLoaded += DeferredSpriteLoadPlaylist_SpriteLoaded;
+                _ = playlist.coverImage;
             }
+
             childPlaylists.Add(playlist);
-            customListTableData.tableView.ReloadData();
-            PlaylistLibUtils.playlistManager.RequestRefresh("PlaylistManager (plugin)");
+            playlistTableData.tableView.ReloadDataKeepingPosition();
+        }
+
+        private void CreateFolder(string folderName)
+        {
+            folderName = folderName.Replace("/", "").Replace("\\", "").Replace(".", "");
+            if (!string.IsNullOrEmpty(folderName))
+            {
+                BeatSaberPlaylistsLib.PlaylistManager childManager = parentManager.CreateChildManager(folderName);
+
+                if (childManagers.Contains(childManager))
+                {
+                    popupModalsController.ShowOkModal(modalTransform, "\"" + folderName + "\" already exists! Please use a different name.", null, animateParentCanvas: false);
+                }
+                else
+                {
+                    playlistTableData.data.Insert(childManagers.Count, new CustomCellInfo(Path.GetFileName(childManager.PlaylistPath), "Folder", folderIcon));
+                    playlistTableData.tableView.ReloadDataKeepingPosition();
+                    childManagers.Add(childManager);
+                    PlaylistLibUtils.playlistManager.RequestRefresh("PlaylistManager (plugin)");
+                }
+            }
         }
 
         #endregion
-
     }
 }
