@@ -25,7 +25,6 @@ namespace PlaylistManager.UI
         private readonly LevelSelectionNavigationController levelSelectionNavigationController;
         private readonly PopupModalsController popupModalsController;
         private readonly HoverHintController hoverHintController;
-        private readonly IVRPlatformHelper platformHelper;
         private BeatmapLevelsModel beatmapLevelsModel;
 
         private FloatingScreen floatingScreen;
@@ -36,10 +35,21 @@ namespace PlaylistManager.UI
 
         public event PropertyChangedEventHandler PropertyChanged;
         public event Action<IAnnotatedBeatmapLevelCollection[], int> LevelCollectionTableViewUpdatedEvent;
+        public event Action<BeatSaberPlaylistsLib.PlaylistManager> ParentManagerUpdatedEvent;
 
-        private BeatSaberPlaylistsLib.PlaylistManager currentParentManager;
+        private BeatSaberPlaylistsLib.PlaylistManager _currentParentManager;
         private List<BeatSaberPlaylistsLib.PlaylistManager> currentManagers;
         private FolderMode folderMode;
+
+        public BeatSaberPlaylistsLib.PlaylistManager CurrentParentManager
+        {
+            get => _currentParentManager;
+            private set
+            {
+                _currentParentManager = value;
+                ParentManagerUpdatedEvent?.Invoke(value);
+            }
+        }
 
         [UIComponent("root")]
         private RectTransform rootTransform;
@@ -58,14 +68,13 @@ namespace PlaylistManager.UI
 
         public FoldersViewController(AnnotatedBeatmapLevelCollectionsViewController annotatedBeatmapLevelCollectionsViewController, MainFlowCoordinator mainFlowCoordinator, 
             LevelSelectionNavigationController levelSelectionNavigationController, PopupModalsController popupModalsController, HoverHintController hoverHintController,
-            IVRPlatformHelper platformHelper, BeatmapLevelsModel beatmapLevelsModel)
+            BeatmapLevelsModel beatmapLevelsModel)
         {
             this.annotatedBeatmapLevelCollectionsViewController = annotatedBeatmapLevelCollectionsViewController;
             this.mainFlowCoordinator = mainFlowCoordinator;
             this.levelSelectionNavigationController = levelSelectionNavigationController;
             this.popupModalsController = popupModalsController;
             this.hoverHintController = hoverHintController;
-            this.platformHelper = platformHelper;
             this.beatmapLevelsModel = beatmapLevelsModel;
 
             levelPacksIcon = BeatSaberMarkupLanguage.Utilities.FindSpriteInAssembly("PlaylistManager.Icons.LevelPacks.png");
@@ -98,7 +107,6 @@ namespace PlaylistManager.UI
             rootTransform.gameObject.name = "PlaylistManagerFoldersView";
 
             ScrollView scrollView = customListTableData.tableView.GetComponent<ScrollView>();
-            Accessors.PlatformHelperAccessor(ref scrollView) = platformHelper;
         }
 
         public void SetupDimensions()
@@ -123,7 +131,7 @@ namespace PlaylistManager.UI
         {
             customListTableData.tableView.ClearSelection();
             customListTableData.data.Clear();
-            this.currentParentManager = currentParentManager;
+            this.CurrentParentManager = currentParentManager;
 
             if (currentParentManager == null)
             {
@@ -222,7 +230,7 @@ namespace PlaylistManager.UI
         [UIAction("folder-select")]
         private void Select(TableView _, int selectedCellIndex)
         {
-            if (currentParentManager == null) // If we are at root
+            if (CurrentParentManager == null) // If we are at root
             {
                 if (selectedCellIndex == 0)
                 {
@@ -258,11 +266,11 @@ namespace PlaylistManager.UI
         [UIAction("back-button-click")]
         private void BackButtonClicked()
         {
-            if (currentParentManager == null)
+            if (CurrentParentManager == null)
             {
                 return;
             }
-            SetupList(currentParentManager: currentParentManager.Parent);
+            SetupList(currentParentManager: CurrentParentManager.Parent);
         }
 
         #region Create Folder
@@ -278,7 +286,7 @@ namespace PlaylistManager.UI
             folderName = folderName.Replace("/", "").Replace("\\", "").Replace(".", "");
             if (!string.IsNullOrEmpty(folderName))
             {
-                BeatSaberPlaylistsLib.PlaylistManager childManager = currentParentManager.CreateChildManager(folderName);
+                BeatSaberPlaylistsLib.PlaylistManager childManager = CurrentParentManager.CreateChildManager(folderName);
 
                 if (currentManagers.Contains(childManager))
                 {
@@ -302,7 +310,7 @@ namespace PlaylistManager.UI
         [UIAction("rename-folder")]
         private void RenameButtonClicked()
         {
-            popupModalsController.ShowKeyboard(levelSelectionNavigationController.transform, RenameKeyboardEnter, keyboardText: Path.GetFileName(currentParentManager.PlaylistPath));
+            popupModalsController.ShowKeyboard(levelSelectionNavigationController.transform, RenameKeyboardEnter, keyboardText: Path.GetFileName(CurrentParentManager.PlaylistPath));
         }
 
         private void RenameKeyboardEnter(string folderName)
@@ -310,9 +318,9 @@ namespace PlaylistManager.UI
             folderName = folderName.Replace("/", "").Replace("\\", "").Replace(".", "");
             if (!string.IsNullOrEmpty(folderName))
             {
-                if (folderName != Path.GetFileName(currentParentManager.PlaylistPath))
+                if (folderName != Path.GetFileName(CurrentParentManager.PlaylistPath))
                 {
-                    currentParentManager.RenameManager(folderName);
+                    CurrentParentManager.RenameManager(folderName);
                     PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(nameof(FolderText)));
                 }
             }
@@ -325,12 +333,12 @@ namespace PlaylistManager.UI
         [UIAction("delete-folder")]
         private void DeleteButtonClicked()
         {
-            popupModalsController.ShowYesNoModal(levelSelectionNavigationController.transform, string.Format("Are you sure you want to delete {0} along with all playlists and subfolders?", Path.GetFileName(currentParentManager.PlaylistPath)), DeleteConfirm);
+            popupModalsController.ShowYesNoModal(levelSelectionNavigationController.transform, string.Format("Are you sure you want to delete {0} along with all playlists and subfolders?", Path.GetFileName(CurrentParentManager.PlaylistPath)), DeleteConfirm);
         }
 
         private void DeleteConfirm()
         {
-            currentParentManager.Parent.DeleteChildManager(currentParentManager);
+            CurrentParentManager.Parent.DeleteChildManager(CurrentParentManager);
             BackButtonClicked();
         }
 
@@ -344,13 +352,12 @@ namespace PlaylistManager.UI
                 SetupDimensions();
                 if (viewControllerActivated)
                 {
-                    SetupList(currentParentManager, false);
+                    SetupList(CurrentParentManager, false);
                 }
             }
             else
             {
                 rootTransform.gameObject.SetActive(false);
-                folderMode = FolderMode.None;
             }
         }
 
@@ -362,6 +369,11 @@ namespace PlaylistManager.UI
 
         public void Refresh()
         {
+            if (!rootTransform.gameObject.activeInHierarchy)
+            {
+                return;
+            }
+
             if (folderMode == FolderMode.AllPacks)
             {
                 IBeatmapLevelPack[] annotatedBeatmapLevelCollections = Accessors.CustomLevelPackCollectionAccessor(ref beatmapLevelsModel).beatmapLevelPacks.Concat(PlaylistLibUtils.playlistManager.GetAllPlaylists(true)).ToArray();
@@ -382,13 +394,13 @@ namespace PlaylistManager.UI
             }
             else if (folderMode == FolderMode.Folders)
             {
-                BeatSaberPlaylistsLib.Types.IPlaylist[] annotatedBeatmapLevelCollections = currentParentManager.GetAllPlaylists(false);
+                BeatSaberPlaylistsLib.Types.IPlaylist[] annotatedBeatmapLevelCollections = CurrentParentManager.GetAllPlaylists(false);
                 int indexToSelect = annotatedBeatmapLevelCollections.IndexOf(annotatedBeatmapLevelCollectionsViewController.selectedAnnotatedBeatmapLevelCollection);
                 if (indexToSelect != -1)
                 {
                     annotatedBeatmapLevelCollectionsViewController.SetData(annotatedBeatmapLevelCollections, indexToSelect, false);
                 }
-                SetupList(currentParentManager, false);
+                SetupList(CurrentParentManager, false);
             }
         }
 
@@ -397,13 +409,13 @@ namespace PlaylistManager.UI
         {
             get
             {
-                if (currentParentManager == null || !Directory.Exists(currentParentManager.PlaylistPath))
+                if (CurrentParentManager == null || !Directory.Exists(CurrentParentManager.PlaylistPath))
                 {
                     return "";
                 }
                 else
                 {
-                    string folderName = Path.GetFileName(currentParentManager.PlaylistPath);
+                    string folderName = Path.GetFileName(CurrentParentManager.PlaylistPath);
                     if (folderName.Length > 15)
                     {
                         return folderName.Substring(0, 15) + "...";
