@@ -1,8 +1,12 @@
-﻿using System.Drawing;
+﻿using System.Collections.Generic;
+using System.Drawing;
 using System.Drawing.Imaging;
 using System.IO;
+using System.Linq;
 using System.Reflection;
 using BeatSaberPlaylistsLib;
+using BeatSaberPlaylistsLib.Blist;
+using BeatSaberPlaylistsLib.Legacy;
 using BeatSaberPlaylistsLib.Types;
 using PlaylistManager.Configuration;
 using UnityEngine;
@@ -11,7 +15,8 @@ namespace PlaylistManager.Utilities
 {
     public class PlaylistLibUtils
     {
-        private static readonly string EASTER_EGG_URL = "https://raw.githubusercontent.com/rithik-b/PlaylistManager/master/img/easteregg.bplist";
+        private const string ICON_PATH = "PlaylistManager.Icons.DefaultIcon.png";
+        private const string EASTER_EGG_URL = "https://raw.githubusercontent.com/rithik-b/PlaylistManager/master/img/easteregg.bplist";
 
         public static BeatSaberPlaylistsLib.PlaylistManager playlistManager
         {
@@ -21,30 +26,39 @@ namespace PlaylistManager.Utilities
             }
         }
 
-        public static BeatSaberPlaylistsLib.Types.IPlaylist CreatePlaylist(string playlistName, string playlistAuthorName, BeatSaberPlaylistsLib.PlaylistManager playlistManager, bool defaultCover = true)
+        public static BeatSaberPlaylistsLib.Types.IPlaylist CreatePlaylistWithConfig(string playlistName, BeatSaberPlaylistsLib.PlaylistManager playlistManager)
+        {
+            string playlistAuthorName = PluginConfig.Instance.AuthorName;
+            bool easterEgg = playlistAuthorName.ToUpper().Contains("BINTER") && playlistName.ToUpper().Contains("TECH") && PluginConfig.Instance.EasterEggs;
+            return CreatePlaylist(playlistName, playlistAuthorName, playlistManager, !PluginConfig.Instance.DefaultImageDisabled, PluginConfig.Instance.DefaultAllowDuplicates, easterEgg);
+        }
+
+        public static BeatSaberPlaylistsLib.Types.IPlaylist CreatePlaylist(string playlistName, string playlistAuthorName, BeatSaberPlaylistsLib.PlaylistManager playlistManager, bool defaultCover = true,
+            bool allowDups = true, bool easterEgg = false)
         {
             BeatSaberPlaylistsLib.Types.IPlaylist playlist = playlistManager.CreatePlaylist("", playlistName, playlistAuthorName, "");
 
             if (defaultCover)
             {
-                using (Stream imageStream = Assembly.GetExecutingAssembly().GetManifestResourceStream("PlaylistManager.Icons.Logo.png"))
+                using (Stream imageStream = Assembly.GetExecutingAssembly().GetManifestResourceStream(ICON_PATH))
                 {
                     playlist.SetCover(imageStream);
                 }
             }
 
-            if (!PluginConfig.Instance.DefaultAllowDuplicates)
+
+            if (!allowDups)
             {
                 playlist.AllowDuplicates = false;
             }
 
-            // Easter Egg
-            if (PluginConfig.Instance.AuthorName.ToUpper().Contains("BINTER") && playlistName.ToUpper().Contains("TECH") && PluginConfig.Instance.EasterEggs)
+            if (easterEgg)
             {
                 playlist.SetCustomData("syncURL", EASTER_EGG_URL);
             }
 
             playlistManager.StorePlaylist(playlist);
+            PlaylistLibUtils.playlistManager.RequestRefresh("PlaylistManager (plugin)");
             return playlist;
         }
 
@@ -63,6 +77,22 @@ namespace PlaylistManager.Utilities
                 return playlistSong.LevelId;
             }
             return "";
+        }
+
+        public static List<IPlaylistSong> GetMissingSongs(BeatSaberPlaylistsLib.Types.IPlaylist playlist, HashSet<string> ownedHashes = null)
+        {
+            if (playlist is LegacyPlaylist legacyPlaylist)
+            {
+                return legacyPlaylist.Where(s => s.PreviewBeatmapLevel == null && !(ownedHashes?.Contains(s.Hash) ?? false)).Distinct(IPlaylistSongComparer<IPlaylistSong>.Default).ToList();
+            }
+            else if (playlist is BlistPlaylist blistPlaylist)
+            {
+                return blistPlaylist.Where(s => s.PreviewBeatmapLevel == null && !(ownedHashes?.Contains(s.Hash) ?? false)).Distinct(IPlaylistSongComparer<IPlaylistSong>.Default).ToList();
+            }
+            else
+            {
+                return null;
+            }
         }
 
         #region Image
